Move React components to own directory
