+++ app/javascript/packs/application.js
@@ -6,7 +6,7 @@
 //
 // To reference this file, add <%= javascript_pack_tag 'application' %> to the appropriate
 // layout file, like app/views/layouts/application.html.erb
-import Finder from './finder'
+import Finder from './components/finder'
 import WebpackerReact from 'webpacker-react'
 
 WebpackerReact.setup({Finder})
