+++ app/javascript/packs/components/stockitems.jsx
